third_party/blink/web_tests/fast/canvas/composited-canvas-toggle-acceleration-expected.txt

CONSOLE WARNING: Canvas2D: Multiple readback operations using getImageData are faster with the willReadFrequently attribute set to true. See: https://html.spec.whatwg.org/multipage/canvas.html#concept-canvas-will-read-frequently
